refactor: 폴더구조 변경 및 interface 컨벤션 통일
