+++ src/api/shop/entity.ts
@@ -1,28 +1,28 @@
-export type ShopDetailInfoParams = {
+export interface ShopDetailInfoParams {
   orderableShopId: number;
-};
+}
 
-type DeliveryTips = {
+interface DeliveryTips {
   from_amount: number;
   to_amount: number;
   fee: number;
-};
+}
 
-type OwnerInfo = {
+interface OwnerInfo {
   name: string;
   shop_name: string;
   address: string;
   company_registration_number: string;
-};
+}
 
-type OriginInfo = {
+interface OriginInfo {
   ingredient: string;
   origin: string;
-};
+}
 
 export type Day = 'MONDAY' | 'TUESDAY' | 'WEDNESDAY' | 'THURSDAY' | 'FRIDAY' | 'SATURDAY' | 'SUNDAY';
 
-export type ShopDetailInfoResponse = {
+export interface ShopDetailInfoResponse {
   shop_id: number;
   orderable_shop_id: number;
   name: string;
@@ -36,7 +36,7 @@ export type ShopDetailInfoResponse = {
   delivery_tips: DeliveryTips[];
   owner_info: OwnerInfo;
   origins: OriginInfo[];
-};
+}
 
 export interface shopDeliveryInfoResponse {
   campus_delivery: boolean;
