+++ src/pages/Delivery/Outside/DetailAddress.tsx
@@ -13,7 +13,7 @@ import BottomModal, {
 } from '@/components/UI/BottomModal/BottomModal';
 import Button from '@/components/UI/Button';
 import Modal, { ModalContent } from '@/components/UI/CenterModal/Modal';
-import { useOrderStore } from '@/store/useOrderStore';
+import { useOrderStore } from '@/stores/useOrderStore';
 import useBooleanState from '@/util/hooks/useBooleanState';
 
 const DetailRequest: string[] = [
