+++ src/pages/Payment/hooks/useConfirmPayments.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
 import { confirmPayments } from '@/api/payments';
-import { ConfirmPaymentsRequest } from '@/types/api/payments';
+import { ConfirmPaymentsRequest } from '@/api/payments/entity';
 
 export default function useConfirmPayments() {
   return useMutation({
