+++ src/pages/Payment/hooks/useTemporaryDelivery.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
 import { getTemporaryDelivery } from '@/api/payments';
-import { DeliveryTemporaryRequest } from '@/types/api/payments';
+import { DeliveryTemporaryRequest } from '@/api/payments/entity';
 
 export default function useTemporaryDelivery() {
   return useMutation({
